+++ src/Instance/ItemsList/ItemsListContainer.js
@@ -15,6 +15,7 @@ import useHoldingItemsQuery from '../../hooks/useHoldingItemsQuery';
 import { DEFAULT_ITEM_TABLE_SORTBY_FIELD } from '../../constants';
 
 const ItemsListContainer = ({
+  tenantId,
   holding,
   draggable,
   droppable,
@@ -36,8 +37,8 @@ const ItemsListContainer = ({
     offset,
   };
 
-  const { isFetching, items } = useHoldingItemsQuery(holding.id, { searchParams });
-  const { totalRecords } = useHoldingItemsQuery(holding.id, { searchParams: { limit: 0 }, key: 'itemCount' });
+  const { isFetching, items } = useHoldingItemsQuery(holding.id, { searchParams, tenantId });
+  const { totalRecords } = useHoldingItemsQuery(holding.id, { searchParams: { limit: 0 }, key: 'itemCount', tenantId });
 
   useEffect(() => {
     if (!isEmpty(items)) {
@@ -68,7 +69,8 @@ const ItemsListContainer = ({
 ItemsListContainer.propTypes = {
   holding: PropTypes.object.isRequired,
   draggable: PropTypes.bool,
-  droppable: PropTypes.bool
+  droppable: PropTypes.bool,
+  tenantId: PropTypes.string,
 };
 
 export default memo(ItemsListContainer);